On branch master  Your branch is up to date with 'origin/master'. 	modified:   databases/createTable.py 	deleted:    databases/getParameter.py 	new file:   databases/insertElement.py 	modified:   databases/parameters.db 	new file:   "envControl/..\\..\\databases\\parameters.db" 	new file:   "envControl/..\\databases\\parameter.db" 	new file:   "envControl/..\\databases\\parameters.db" 	new file:   "envControl/\\..\\..\\databases\\parameters.db" 	new file:   envControl/control/__init__.py 	modified:   envControl/control/airControl.py 	new file:   envControl/control/dataBase.py 	modified:   envControl/control/pHControl.py 	modified:   envControl/control/sensorControl.py 	modified:   envControl/control/serialBase.py 	new file:   "envControl/home\\pi\\grow\\databases\\parameter.db" 	new file:   "envControl/home\\pi\\grow\\databases\\parameters.db" 	modified:   envControl/manualControl.py 	new file:   envControl/tests/__init__.py 	modified:   envControl/tests/pHcalibrate.py

+++ databases/createTable.py
@@ -6,7 +6,7 @@ print("Opened database successfully")
 conn.execute('''CREATE TABLE CALIBRATION
 		(ID INTEGER PRIMARY KEY AUTOINCREMENT,
 		TYPE     TEXT   NOT NULL,
-		VALUE        INT    NOT NULL);''')
+		VALUE        FLOAT    NOT NULL);''')
 
 print("Table created successfuly")
